@@ -31,7 +31,7 @@ class ProductTest extends UnitTestCase
      *
      * @return array
      */
-    public function bestSpecialPriceProvider()
+    public static function bestSpecialPriceProvider()
     {
         return [
             [100.0, 80.0, 75.0, 90.0, 75.0],
@@ -45,7 +45,7 @@ class ProductTest extends UnitTestCase
      *
      * @return array
      */
-    public function bestSpecialPriceDiscountProvider()
+    public static function bestSpecialPriceDiscountProvider()
     {
         return [
             [100.0, 80.0, 75.0, 90.0, 25.0],
@@ -57,9 +57,9 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getProductTypeReturnsInitialValueForProductType()
+    public function getProductTypeReturnsInitialValueForProductType(): void
     {
-        $this->assertSame(
+        self::assertSame(
             'simple',
             $this->product->getProductType()
         );
@@ -68,11 +68,11 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setProductTypeSetsProductType()
+    public function setProductTypeSetsProductType(): void
     {
         $this->product->setProductType('configurable');
 
-        $this->assertSame(
+        self::assertSame(
             'configurable',
             $this->product->getProductType()
         );
@@ -81,9 +81,9 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getTeaserReturnsInitialValueForTeaser()
+    public function getTeaserReturnsInitialValueForTeaser(): void
     {
-        $this->assertSame(
+        self::assertSame(
             '',
             $this->product->getTeaser()
         );
@@ -92,11 +92,11 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setTeaserForStringSetsTeaser()
+    public function setTeaserForStringSetsTeaser(): void
     {
         $this->product->setTeaser('Conceived at T3CON10');
 
-        $this->assertSame(
+        self::assertSame(
             'Conceived at T3CON10',
             $this->product->getTeaser()
         );
@@ -105,9 +105,9 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getMinNumberInOrderInitiallyReturnsMinNumberInOrder()
+    public function getMinNumberInOrderInitiallyReturnsMinNumberInOrder(): void
     {
-        $this->assertSame(
+        self::assertSame(
             0,
             $this->product->getMinNumberInOrder()
         );
@@ -116,7 +116,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setNegativeMinNumberThrowsException()
+    public function setNegativeMinNumberThrowsException(): void
     {
         $this->expectException(\InvalidArgumentException::class);
         $minNumber = -10;
@@ -127,7 +127,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setMinNumberGreaterThanMaxNumberThrowsException()
+    public function setMinNumberGreaterThanMaxNumberThrowsException(): void
     {
         $this->expectException(\InvalidArgumentException::class);
         $minNumber = 10;
@@ -138,14 +138,14 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setMinNumberInOrderSetsMinNumberInOrder()
+    public function setMinNumberInOrderSetsMinNumberInOrder(): void
     {
         $minNumber = 10;
 
         $this->product->setMaxNumberInOrder($minNumber);
         $this->product->setMinNumberInOrder($minNumber);
 
-        $this->assertSame(
+        self::assertSame(
             $minNumber,
             $this->product->getMinNumberInOrder()
         );
@@ -154,9 +154,9 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getMaxNumberInOrderInitiallyReturnsMaxNumberInOrder()
+    public function getMaxNumberInOrderInitiallyReturnsMaxNumberInOrder(): void
     {
-        $this->assertSame(
+        self::assertSame(
             0,
             $this->product->getMaxNumberInOrder()
         );
@@ -165,7 +165,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setNegativeMaxNumberThrowsException()
+    public function setNegativeMaxNumberThrowsException(): void
     {
         $this->expectException(\InvalidArgumentException::class);
         $maxNumber = -10;
@@ -176,13 +176,13 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setMaxNumberInOrderSetsMaxNumberInOrder()
+    public function setMaxNumberInOrderSetsMaxNumberInOrder(): void
     {
         $maxNumber = 10;
 
         $this->product->setMaxNumberInOrder($maxNumber);
 
-        $this->assertSame(
+        self::assertSame(
             $maxNumber,
             $this->product->getMaxNumberInOrder()
         );
@@ -191,7 +191,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setMaxNumberLesserThanMinNumberThrowsException()
+    public function setMaxNumberLesserThanMinNumberThrowsException(): void
     {
         $this->expectException(\InvalidArgumentException::class);
         $minNumber = 10;
@@ -206,9 +206,9 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getPriceReturnsInitialValueForFloat()
+    public function getPriceReturnsInitialValueForFloat(): void
     {
-        $this->assertSame(
+        self::assertSame(
             0.0,
             $this->product->getPrice()
         );
@@ -217,11 +217,11 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setPriceSetsPrice()
+    public function setPriceSetsPrice(): void
     {
         $this->product->setPrice(3.14159265);
 
-        $this->assertSame(
+        self::assertSame(
             3.14159265,
             $this->product->getPrice()
         );
@@ -230,9 +230,9 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getSpecialPricesInitiallyIsEmpty()
+    public function getSpecialPricesInitiallyIsEmpty(): void
     {
-        $this->assertEmpty(
+        self::assertEmpty(
             $this->product->getSpecialPrices()
         );
     }
@@ -240,7 +240,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setSpecialPricesSetsSpecialPrices()
+    public function setSpecialPricesSetsSpecialPrices(): void
     {
         $price = 10.00;
 
@@ -252,7 +252,7 @@ class ProductTest extends UnitTestCase
 
         $this->product->setSpecialPrices($objectStorage);
 
-        $this->assertContains(
+        self::assertContains(
             $specialPrice,
             $this->product->getSpecialPrices()
         );
@@ -261,7 +261,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function addSpecialPriceAddsSpecialPrice()
+    public function addSpecialPriceAddsSpecialPrice(): void
     {
         $price = 10.00;
 
@@ -270,7 +270,7 @@ class ProductTest extends UnitTestCase
 
         $this->product->addSpecialPrice($specialPrice);
 
-        $this->assertContains(
+        self::assertContains(
             $specialPrice,
             $this->product->getSpecialPrices()
         );
@@ -279,7 +279,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function removeSpecialPriceRemovesSpecialPrice()
+    public function removeSpecialPriceRemovesSpecialPrice(): void
     {
         $price = 10.00;
 
@@ -289,7 +289,7 @@ class ProductTest extends UnitTestCase
         $this->product->addSpecialPrice($specialPrice);
         $this->product->removeSpecialPrice($specialPrice);
 
-        $this->assertEmpty(
+        self::assertEmpty(
             $this->product->getSpecialPrices()
         );
     }
@@ -297,14 +297,14 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getBestSpecialPriceDiscountForEmptySpecialPriceReturnsDiscount()
+    public function getBestSpecialPriceDiscountForEmptySpecialPriceReturnsDiscount(): void
     {
         $price = 10.00;
 
         $product = new Product();
         $product->setPrice($price);
 
-        $this->assertSame(
+        self::assertSame(
             0.0,
             $product->getBestSpecialPriceDiscount()
         );
@@ -320,7 +320,7 @@ class ProductTest extends UnitTestCase
         $special2,
         $special3,
         $expectedBestSpecialPrice
-    ) {
+    ): void {
         $product = new Product();
         $product->setPrice($price);
 
@@ -336,7 +336,7 @@ class ProductTest extends UnitTestCase
         $specialPrice3->setPrice($special3);
         $product->addSpecialPrice($specialPrice3);
 
-        $this->assertSame(
+        self::assertSame(
             $expectedBestSpecialPrice,
             $product->getBestSpecialPrice()
         );
@@ -345,7 +345,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getBestSpecialPriceDiscountForGivenSpecialPriceReturnsPercentageDiscount()
+    public function getBestSpecialPriceDiscountForGivenSpecialPriceReturnsPercentageDiscount(): void
     {
         $price = 10.0;
         $porductSpecialPrice = 9.0;
@@ -358,7 +358,7 @@ class ProductTest extends UnitTestCase
 
         $product->addSpecialPrice($specialPrice);
 
-        $this->assertSame(
+        self::assertSame(
             10.0,
             $product->getBestSpecialPricePercentageDiscount()
         );
@@ -374,7 +374,7 @@ class ProductTest extends UnitTestCase
         $special2,
         $special3,
         $expectedBestSpecialPriceDiscount
-    ) {
+    ): void {
         $product = new Product();
         $product->setPrice($price);
 
@@ -390,7 +390,7 @@ class ProductTest extends UnitTestCase
         $specialPrice3->setPrice($special3);
         $product->addSpecialPrice($specialPrice3);
 
-        $this->assertSame(
+        self::assertSame(
             $expectedBestSpecialPriceDiscount,
             $product->getBestSpecialPriceDiscount()
         );
@@ -399,11 +399,11 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getStockWithoutHandleStockInitiallyReturnsIntMax()
+    public function getStockWithoutHandleStockInitiallyReturnsIntMax(): void
     {
         $product = new Product();
 
-        $this->assertSame(
+        self::assertSame(
             PHP_INT_MAX,
             $product->getStock()
         );
@@ -412,12 +412,12 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getStockWithHandleStockInitiallyReturnsZero()
+    public function getStockWithHandleStockInitiallyReturnsZero(): void
     {
         $product = new Product();
         $product->setHandleStock(true);
 
-        $this->assertSame(
+        self::assertSame(
             0,
             $product->getStock()
         );
@@ -426,7 +426,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setStockWithHandleStockSetsStock()
+    public function setStockWithHandleStockSetsStock(): void
     {
         $stock = 10;
 
@@ -434,14 +434,14 @@ class ProductTest extends UnitTestCase
         $product->setStock($stock);
         $product->setHandleStock(true);
 
-        $this->assertSame(
+        self::assertSame(
             $stock,
             $product->getStock()
         );
 
         $product->setHandleStock(false);
 
-        $this->assertSame(
+        self::assertSame(
             PHP_INT_MAX,
             $product->getStock()
         );
@@ -450,7 +450,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function addToStockAddsANumberOfProductsToStock()
+    public function addToStockAddsANumberOfProductsToStock(): void
     {
         $numberOfProducts = 10;
 
@@ -458,7 +458,7 @@ class ProductTest extends UnitTestCase
         $product->setHandleStock(true);
         $product->addToStock($numberOfProducts);
 
-        $this->assertSame(
+        self::assertSame(
             $numberOfProducts,
             $product->getStock()
         );
@@ -467,7 +467,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function removeFromStockAddsRemovesANumberOfProductsFromStock()
+    public function removeFromStockAddsRemovesANumberOfProductsFromStock(): void
     {
         $stock = 100;
         $numberOfProducts = 10;
@@ -477,7 +477,7 @@ class ProductTest extends UnitTestCase
         $product->setStock($stock);
         $product->removeFromStock($numberOfProducts);
 
-        $this->assertSame(
+        self::assertSame(
             ($stock - $numberOfProducts),
             $product->getStock()
         );
@@ -486,11 +486,11 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function handleStockInitiallyReturnsFalse()
+    public function handleStockInitiallyReturnsFalse(): void
     {
         $product = new Product();
 
-        $this->assertFalse(
+        self::assertFalse(
             $product->isHandleStock()
         );
     }
@@ -498,12 +498,12 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setHandleStockSetsHandleStock()
+    public function setHandleStockSetsHandleStock(): void
     {
         $product = new Product();
         $product->setHandleStock(true);
 
-        $this->assertTrue(
+        self::assertTrue(
             $product->isHandleStock()
         );
     }
@@ -511,11 +511,11 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function isAvailableInitiallyReturnsTrue()
+    public function isAvailableInitiallyReturnsTrue(): void
     {
         $product = new Product();
 
-        $this->assertTrue(
+        self::assertTrue(
             $product->getIsAvailable()
         );
     }
@@ -523,12 +523,12 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function isAvailableWithHandleStockIsEnabledAndEmptyStockReturnsFalse()
+    public function isAvailableWithHandleStockIsEnabledAndEmptyStockReturnsFalse(): void
     {
         $product = new Product();
         $product->setHandleStock(true);
 
-        $this->assertFalse(
+        self::assertFalse(
             $product->getIsAvailable()
         );
     }
@@ -536,13 +536,13 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function isAvailableWithHandleStockIsEnabledAndNotEmptyStockReturnsTrue()
+    public function isAvailableWithHandleStockIsEnabledAndNotEmptyStockReturnsTrue(): void
     {
         $product = new Product();
         $product->setStock(10);
         $product->setHandleStock(true);
 
-        $this->assertTrue(
+        self::assertTrue(
             $product->getIsAvailable()
         );
     }
@@ -550,14 +550,14 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function isAvailableWithHandleStockAndHandleStockInVariantsIsEnabledAndNoBackendVariantsConfiguredReturnsFalse()
+    public function isAvailableWithHandleStockAndHandleStockInVariantsIsEnabledAndNoBackendVariantsConfiguredReturnsFalse(): void
     {
         $product = new Product();
         $product->setStock(10);
         $product->setHandleStock(true);
         $product->setHandleStockInVariants(true);
 
-        $this->assertFalse(
+        self::assertFalse(
             $product->getIsAvailable()
         );
     }
@@ -565,12 +565,12 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function isAvailableWithHandleStockAndHandleStockInVariantsIsEnabledAndBackendVariantConfiguredIsNotAvailableReturnsFalse()
+    public function isAvailableWithHandleStockAndHandleStockInVariantsIsEnabledAndBackendVariantConfiguredIsNotAvailableReturnsFalse(): void
     {
         $productBackendVariant = $this->createMock(
             BeVariant::class
         );
-        $productBackendVariant->expects($this->any())->method('getIsAvailable')->will($this->returnValue(false));
+        $productBackendVariant->expects(self::any())->method('getIsAvailable')->willReturn(false);
 
         $product = new Product();
         $product->addBeVariant($productBackendVariant);
@@ -578,7 +578,7 @@ class ProductTest extends UnitTestCase
         $product->setHandleStock(true);
         $product->setHandleStockInVariants(true);
 
-        $this->assertFalse(
+        self::assertFalse(
             $product->getIsAvailable()
         );
     }
@@ -586,12 +586,12 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function isAvailableWithHandleStockAndHandleStockInVariantsIsEnabledAndBackendVariantConfiguredIsAvailableReturnsFalse()
+    public function isAvailableWithHandleStockAndHandleStockInVariantsIsEnabledAndBackendVariantConfiguredIsAvailableReturnsFalse(): void
     {
         $productBackendVariant = $this->createMock(
             BeVariant::class
         );
-        $productBackendVariant->expects($this->any())->method('getIsAvailable')->will($this->returnValue(true));
+        $productBackendVariant->expects(self::any())->method('getIsAvailable')->willReturn(true);
 
         $product = new Product();
         $product->addBeVariant($productBackendVariant);
@@ -599,7 +599,7 @@ class ProductTest extends UnitTestCase
         $product->setHandleStock(true);
         $product->setHandleStockInVariants(true);
 
-        $this->assertTrue(
+        self::assertTrue(
             $product->getIsAvailable()
         );
     }
@@ -607,290 +607,9 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getPriceMeasure()
+    public function getTaxClassIdInitiallyReturnsTaxClassId(): void
     {
-        $this->assertSame(
-            0.0,
-            $this->product->getPriceMeasure()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function setPriceMeasureSetsPriceMeasure()
-    {
-        $priceMeasure = 10.99;
-
-        $this->product->setPriceMeasure($priceMeasure);
-
-        $this->assertSame(
-            $priceMeasure,
-            $this->product->getPriceMeasure()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getPriceMeasureUnit()
-    {
-        $this->assertSame(
-            '',
-            $this->product->getPriceMeasureUnit()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function setPriceMeasureUnitSetsPriceMeasureUnit()
-    {
-        $priceMeasureUnit = 'l';
-
-        $this->product->setPriceMeasureUnit($priceMeasureUnit);
-
-        $this->assertSame(
-            $priceMeasureUnit,
-            $this->product->getPriceMeasureUnit()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getBasePriceMeasureUnit()
-    {
-        $this->assertSame(
-            '',
-            $this->product->getBasePriceMeasureUnit()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function setBasePriceMeasureUnitSetsBasePriceMeasureUnit()
-    {
-        $priceBaseMeasureUnit = 'l';
-
-        $this->product->setBasePriceMeasureUnit($priceBaseMeasureUnit);
-
-        $this->assertSame(
-            $priceBaseMeasureUnit,
-            $this->product->getBasePriceMeasureUnit()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getIsMeasureUnitCompatibilityInitiallyRetrunsFalse()
-    {
-        $product = new Product();
-
-        $this->assertFalse(
-            $product->getIsMeasureUnitCompatibility()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getIsMeasureUnitCompatibilityAndNotSetPriceMeasureUnitsRetrunsFalse()
-    {
-        $product = new Product();
-        $product->setBasePriceMeasureUnit('l');
-
-        $this->assertFalse(
-            $product->getIsMeasureUnitCompatibility()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getIsMeasureUnitCompatibilityAndNotSetBasePriceMeasureUnitsRetrunsFalse()
-    {
-        $product = new Product();
-        $product->setPriceMeasureUnit('l');
-
-        $this->assertFalse(
-            $product->getIsMeasureUnitCompatibility()
-        );
-    }
-
-    /**
-     * Measurement Units Provider
-     *
-     * @return array
-     */
-    public function measureUnitsProvider()
-    {
-        return [
-            ['mg',  'kg', 1000000.0, 1000.0, 1000.0],
-            ['g',   'kg', 1000.0,    1000.0, 1.0],
-            ['kg',  'kg', 1.0,       1000.0, 0.001],
-            ['ml',  'l',  1000.0,    1000.0, 1.0],
-            ['cl',  'l',  100.0,     1000.0, 0.1],
-            ['l',   'l',  1.0,       1000.0, 0.001],
-            ['cbm', 'l',  0.001,     1.0,    0.001],
-            ['mm',  'm',  1000.0,    1000.0, 1.0],
-            ['cm',  'm',  100.0,     1000.0, 0.1],
-            ['m',   'm',  1.0,       2.0,    0.5],
-            ['km',  'm',  0.001,     2.0,    0.0005],
-            ['m2',  'm2', 1.0,       20.0,   0.05],
-        ];
-    }
-
-    /**
-     * @test
-     * @dataProvider measureUnitsProvider
-     */
-    public function getIsMeasureUnitCompatibilityRetrunsTrueOnSameTypeOfMeasureUnit(
-        $sourceMeasureUnit,
-        $targetMeasureUnit,
-        $factor,
-        $priceMeasure,
-        $calculatedBasePrice
-    ) {
-        $product = new Product();
-        $product->setPriceMeasureUnit($sourceMeasureUnit);
-        $product->setBasePriceMeasureUnit($targetMeasureUnit);
-
-        $this->assertTrue(
-            $product->getIsMeasureUnitCompatibility()
-        );
-    }
-
-    /**
-     * @test
-     * @dataProvider measureUnitsProvider
-     */
-    public function getMeasureUnitFactorForGivenPriceMeasureUnitAndBasePriceMeasureUnitRetrunsFactor(
-        $sourceMeasureUnit,
-        $targetMeasureUnit,
-        $factor,
-        $priceMeasure,
-        $calculatedBasePrice
-    ) {
-        $product = new Product();
-        $product->setPriceMeasureUnit($sourceMeasureUnit);
-        $product->setBasePriceMeasureUnit($targetMeasureUnit);
-        $product->setPriceMeasure(1);
-
-        $this->assertSame(
-            $factor,
-            $product->getMeasureUnitFactor()
-        );
-    }
-
-    /**
-     * @test
-     * @dataProvider measureUnitsProvider
-     */
-    public function getCalculatedBasePriceForGivenPriceMeasureUnitAndBasePriceMeasureUnitRetrunsPrice(
-        $sourceMeasureUnit,
-        $targetMeasureUnit,
-        $factor,
-        $priceMeasure,
-        $calculatedBasePrice
-    ) {
-        $product = new Product();
-        $product->setPriceMeasureUnit($sourceMeasureUnit);
-        $product->setBasePriceMeasureUnit($targetMeasureUnit);
-        $product->setPriceMeasure($priceMeasure);
-
-        $this->assertSame(
-            $calculatedBasePrice,
-            $product->getMeasureUnitFactor()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getServiceAttribute1ReturnsZero()
-    {
-        $this->assertSame(
-            0.0,
-            $this->product->getServiceAttribute1()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function setServiceAttribute1SetsServiceAttribute1()
-    {
-        $serviceAttribute1 = 1.0;
-
-        $this->product->setServiceAttribute1($serviceAttribute1);
-
-        $this->assertSame(
-            $serviceAttribute1,
-            $this->product->getServiceAttribute1()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getServiceAttribute2ReturnsZero()
-    {
-        $this->assertSame(
-            0.0,
-            $this->product->getServiceAttribute2()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function setServiceAttribute2SetsServiceAttribute2()
-    {
-        $serviceAttribute2 = 2.0;
-
-        $this->product->setServiceAttribute2($serviceAttribute2);
-
-        $this->assertSame(
-            $serviceAttribute2,
-            $this->product->getServiceAttribute2()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getServiceAttribute3ReturnsZero()
-    {
-        $this->assertSame(
-            0.0,
-            $this->product->getServiceAttribute3()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function setServiceAttribute3SetsServiceAttribute3()
-    {
-        $serviceAttribute3 = 3.0;
-
-        $this->product->setServiceAttribute3($serviceAttribute3);
-
-        $this->assertSame(
-            $serviceAttribute3,
-            $this->product->getServiceAttribute3()
-        );
-    }
-
-    /**
-     * @test
-     */
-    public function getTaxClassIdInitiallyReturnsTaxClassId()
-    {
-        $this->assertSame(
+        self::assertSame(
             1,
             $this->product->getTaxClassId()
         );
@@ -899,13 +618,13 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setTaxClassIdSetsTaxClassId()
+    public function setTaxClassIdSetsTaxClassId(): void
     {
         $taxClassId = 2;
 
         $this->product->setTaxClassId($taxClassId);
 
-        $this->assertSame(
+        self::assertSame(
             $taxClassId,
             $this->product->getTaxClassId()
         );
@@ -914,9 +633,9 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getBeVariantAttribute1InitiallyIsNull()
+    public function getBeVariantAttribute1InitiallyIsNull(): void
     {
-        $this->assertNull(
+        self::assertNull(
             $this->product->getBeVariantAttribute1()
         );
     }
@@ -924,13 +643,13 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setBeVariantAttribute1SetsBeVariantAttribute1()
+    public function setBeVariantAttribute1SetsBeVariantAttribute1(): void
     {
         $beVariantAttribute = new BeVariantAttribute();
 
         $this->product->setBeVariantAttribute1($beVariantAttribute);
 
-        $this->assertSame(
+        self::assertSame(
             $beVariantAttribute,
             $this->product->getBeVariantAttribute1()
         );
@@ -939,9 +658,9 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getBeVariantAttribute2InitiallyIsNull()
+    public function getBeVariantAttribute2InitiallyIsNull(): void
     {
-        $this->assertNull(
+        self::assertNull(
             $this->product->getBeVariantAttribute2()
         );
     }
@@ -949,13 +668,13 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setBeVariantAttribute2SetsBeVariantAttribute2()
+    public function setBeVariantAttribute2SetsBeVariantAttribute2(): void
     {
         $beVariantAttribute = new BeVariantAttribute();
 
         $this->product->setBeVariantAttribute2($beVariantAttribute);
 
-        $this->assertSame(
+        self::assertSame(
             $beVariantAttribute,
             $this->product->getBeVariantAttribute2()
         );
@@ -964,9 +683,9 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getBeVariantAttribute3InitiallyIsNull()
+    public function getBeVariantAttribute3InitiallyIsNull(): void
     {
-        $this->assertNull(
+        self::assertNull(
             $this->product->getBeVariantAttribute3()
         );
     }
@@ -974,13 +693,13 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setBeVariantAttribute3SetsBeVariantAttribute3()
+    public function setBeVariantAttribute3SetsBeVariantAttribute3(): void
     {
         $beVariantAttribute = new BeVariantAttribute();
 
         $this->product->setBeVariantAttribute3($beVariantAttribute);
 
-        $this->assertSame(
+        self::assertSame(
             $beVariantAttribute,
             $this->product->getBeVariantAttribute3()
         );
@@ -989,9 +708,9 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function getVariantsInitiallyIsEmpty()
+    public function getVariantsInitiallyIsEmpty(): void
     {
-        $this->assertEmpty(
+        self::assertEmpty(
             $this->product->getBeVariants()
         );
     }
@@ -999,7 +718,7 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function setVariantsSetsVariants()
+    public function setVariantsSetsVariants(): void
     {
         $variant = new BeVariant();
 
@@ -1008,7 +727,7 @@ class ProductTest extends UnitTestCase
 
         $this->product->setBeVariants($objectStorage);
 
-        $this->assertContains(
+        self::assertContains(
             $variant,
             $this->product->getBeVariants()
         );
@@ -1017,13 +736,13 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function addVariantAddsVariant()
+    public function addVariantAddsVariant(): void
     {
         $variant = new BeVariant();
 
         $this->product->addBeVariant($variant);
 
-        $this->assertContains(
+        self::assertContains(
             $variant,
             $this->product->getBeVariants()
         );
@@ -1032,14 +751,14 @@ class ProductTest extends UnitTestCase
     /**
      * @test
      */
-    public function removeVariantRemovesVariant()
+    public function removeVariantRemovesVariant(): void
     {
         $variant = new BeVariant();
 
         $this->product->addBeVariant($variant);
         $this->product->removeBeVariant($variant);
 
-        $this->assertEmpty(
+        self::assertEmpty(
             $this->product->getBeVariants()
         );
     }
