@@ -2,113 +2,52 @@
 
 declare(strict_types=1);
 
-use Rector\Core\Configuration\Option;
-use Rector\Core\ValueObject\PhpVersion;
-use Rector\Php55\Rector\String_\StringClassNameToClassConstantRector;
+use Rector\Config\RectorConfig;
 use Rector\PostRector\Rector\NameImportingPostRector;
+use Rector\TypeDeclaration\Rector\ClassMethod\AddVoidReturnTypeWhereNoReturnRector;
+use Rector\ValueObject\PhpVersion;
+use Ssch\TYPO3Rector\CodeQuality\General\ConvertImplicitVariablesToExplicitGlobalsRector;
+use Ssch\TYPO3Rector\CodeQuality\General\ExtEmConfRector;
 use Ssch\TYPO3Rector\Configuration\Typo3Option;
-use Ssch\TYPO3Rector\FileProcessor\Composer\Rector\ExtensionComposerRector;
-use Ssch\TYPO3Rector\FileProcessor\TypoScript\Visitors\FileIncludeToImportStatementVisitor;
-use Ssch\TYPO3Rector\Rector\General\ConvertTypo3ConfVarsRector;
-use Ssch\TYPO3Rector\Rector\General\ExtEmConfRector;
-use Ssch\TYPO3Rector\Rector\v9\v0\InjectAnnotationRector;
+use Ssch\TYPO3Rector\Set\Typo3LevelSetList;
 use Ssch\TYPO3Rector\Set\Typo3SetList;
-use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
 
-return static function (ContainerConfigurator $containerConfigurator): void {
-    $parameters = $containerConfigurator->parameters();
-
-    $containerConfigurator->import(Typo3SetList::TYPO3_76);
-    $containerConfigurator->import(Typo3SetList::TYPO3_87);
-    $containerConfigurator->import(Typo3SetList::TYPO3_95);
-    $containerConfigurator->import(Typo3SetList::TYPO3_104);
-
-    // In order to have a better analysis from phpstan we teach it here some more things
-    $parameters->set(Option::PHPSTAN_FOR_RECTOR_PATH, Typo3Option::PHPSTAN_FOR_RECTOR_PATH);
-
-    // FQN classes are not imported by default. If you don't do it manually after every Rector run, enable it by:
-    $parameters->set(Option::AUTO_IMPORT_NAMES, true);
-
-    // this will not import root namespace classes, like \DateTime or \Exception
-    $parameters->set(Option::IMPORT_SHORT_CLASSES, false);
-
-    // this will not import classes used in PHP DocBlocks, like in /** @var \Some\Class */
-    $parameters->set(Option::IMPORT_DOC_BLOCKS, false);
-
-    // Define your target version which you want to support
-    $parameters->set(Option::PHP_VERSION_FEATURES, PhpVersion::PHP_72);
-
-    // If you have an editorconfig and changed files should keep their format enable it here
-    // $parameters->set(Option::ENABLE_EDITORCONFIG, true);
-
-    // If you only want to process one/some TYPO3 extension(s), you can specify its path(s) here.
-    // If you use the option --config change __DIR__ to getcwd()
-    // $parameters->set(Option::PATHS, [
-    //    __DIR__ . '/packages/acme_demo/',
-    // ]);
-
-    // If you set option Option::AUTO_IMPORT_NAMES to true, you should consider excluding some TYPO3 files.
-    // If you use the option --config change __DIR__ to getcwd()
-    $parameters->set(Option::SKIP, [
+return RectorConfig::configure()
+    ->withImportNames(true, true, false, true)
+    ->withPaths([
+        __DIR__ . '/Classes',
+        __DIR__ . '/Configuration',
+        __DIR__ . '/Tests',
+        __DIR__ . '/ext_emconf.php',
+        __DIR__ . '/ext_localconf.php',
+    ])
+    // uncomment to reach your current PHP version
+     ->withPhpSets(php81: true)
+    ->withPhpVersion(PhpVersion::PHP_81)
+    ->withSets([
+        Typo3SetList::CODE_QUALITY,
+        Typo3SetList::GENERAL,
+        Typo3LevelSetList::UP_TO_TYPO3_12,
+    ])
+    // To have a better analysis from PHPStan, we teach it here some more things
+    ->withPHPStanConfigs([
+        Typo3Option::PHPSTAN_FOR_RECTOR_PATH,
+    ])
+    ->withRules([
+        AddVoidReturnTypeWhereNoReturnRector::class,
+        ConvertImplicitVariablesToExplicitGlobalsRector::class,
+    ])
+    ->withConfiguredRule(ExtEmConfRector::class, [
+        ExtEmConfRector::PHP_VERSION_CONSTRAINT => '8.1.0-8.3.99',
+        ExtEmConfRector::TYPO3_VERSION_CONSTRAINT => '12.4.0-12.4.99',
+        ExtEmConfRector::ADDITIONAL_VALUES_TO_BE_REMOVED => [],
+    ])
+    // If you use withImportNames(), you should consider excluding some TYPO3 files.
+    ->withSkip([
+        // @see https://github.com/sabbelasichon/typo3-rector/issues/2536
+        __DIR__ . '/**/Configuration/ExtensionBuilder/*',
         NameImportingPostRector::class => [
             'ClassAliasMap.php',
-            'ext_emconf.php',
-            'ext_localconf.php',
-            'ext_tables.php',
-            __DIR__ . '/**/Configuration/AjaxRoutes.php',
-            __DIR__ . '/**/Configuration/Backend/AjaxRoutes.php',
-            __DIR__ . '/**/Configuration/Commands.php',
-            __DIR__ . '/**/Configuration/ExpressionLanguage.php',
-            __DIR__ . '/**/Configuration/Extbase/Persistence/Classes.php',
-            __DIR__ . '/**/Configuration/RequestMiddlewares.php',
-            __DIR__ . '/**/Configuration/TCA/*',
         ],
-
-        // We skip those directories on purpose as there might be node_modules or similar
-        // that include typescript which would result in false positive processing
-        __DIR__ . '/**/Resources/**/node_modules/*',
-        __DIR__ . '/**/Resources/**/NodeModules/*',
-        __DIR__ . '/**/Resources/**/BowerComponents/*',
-        __DIR__ . '/**/Resources/**/bower_components/*',
-        __DIR__ . '/**/Resources/**/build/*',
-    ]);
-
-    // If you have trouble that rector cannot run because some TYPO3 constants are not defined add an additional constants file
-    // @see https://github.com/sabbelasichon/typo3-rector/blob/master/typo3.constants.php
-    // @see https://github.com/rectorphp/rector/blob/main/docs/static_reflection_and_autoload.md#include-files
-    // $parameters->set(Option::BOOTSTRAP_FILES, [
-    //    __DIR__ . '/typo3.constants.php'
-    // ]);
-
-    // get services (needed for register a single rule)
-    $services = $containerConfigurator->services();
-
-    // register a single rule
-    // $services->set(InjectAnnotationRector::class);
-
-    /**
-     * Useful rule from RectorPHP itself to transform i.e. GeneralUtility::makeInstance('TYPO3\CMS\Core\Log\LogManager')
-     * to GeneralUtility::makeInstance(\TYPO3\CMS\Core\Log\LogManager::class) calls.
-     * But be warned, sometimes it produces false positives (edge cases), so watch out
-     */
-    // $services->set(StringClassNameToClassConstantRector::class);
-
-    // Optional non-php file functionalities:
-    // @see https://github.com/sabbelasichon/typo3-rector/blob/main/docs/beyond_php_file_processors.md
-
-    // Adapt your composer.json dependencies to the latest available version for the defined SetList
-    // $containerConfigurator->import(Typo3SetList::COMPOSER_PACKAGES_104_CORE);
-    // $containerConfigurator->import(Typo3SetList::COMPOSER_PACKAGES_104_EXTENSIONS);
-
-    // Rewrite your extbase persistence class mapping from typoscript into php according to official docs.
-    // This processor will create a summarized file with all of the typoscript rewrites combined into a single file.
-    // The filename can be passed as argument, "Configuration_Extbase_Persistence_Classes.php" is default.
-    // $services->set(ExtbasePersistenceVisitor::class);
-    // Add some general TYPO3 rules
-    $services->set(ConvertTypo3ConfVarsRector::class);
-    $services->set(ExtEmConfRector::class);
-    $services->set(ExtensionComposerRector::class);
-
-    // Do you want to modernize your TypoScript include statements for files and move from <INCLUDE /> to @import use the FileIncludeToImportStatementVisitor
-    // $services->set(FileIncludeToImportStatementVisitor::class);
-};
+    ])
+;
